refactor: extract ColumnTitle component
