+++ management-app/src/components/ColumnTitle/ColumnTitle.types.ts
@@ -1,0 +1,6 @@
+import { Dispatch } from 'react';
+
+export type ColumnTitleProps = {
+  title: string;
+  handleClickAway: Dispatch<string>;
+};
