+++ management-app/src/components/ColumnTitle/index.ts
@@ -1,0 +1,2 @@
+export * from './ColumnTitle';
+export * from './ColumnTitle.types';
